pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jaguar_project"
version = "0.1.0"
description = "Jaguars"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "pyyaml>=6.0",
  "scikit-learn>=1.4",
  "tqdm>=4.66",
  "pytest>=8.0",
  "torch>=2.2",
  "scipy>=1.4",
  "matplotlib>=3.7",
  "huggingface-hub>=0.23",
  "pytest>=8.3.3",
  "torchvision==0.24.0",
  "pillow==11.3.0",
  "wandb==0.23.1",
  "python-dotenv==1.2.1",
  "fiftyone==1.11.0",
  "grad-cam==1.5.5",
  "transformers==4.51.0",
  "faiss-cpu==1.13.2",
  "timm",
  "quantus==0.6.0",
  "pyarrow==23.0.1",
  "ImageHash",
  "seaborn==0.13.2"
]

[project.optional-dependencies]
dev = [
  "pytest >=8.3.3", 
]

[tool.setuptools]
package-dir = {"" = "src"}
